Improve UI class organisation by feature packages
Improve black/white list fragments by refactoring code into an abstract list fragment
