+++ AdAway/src/main/java/org/adaway/ui/MainActivity.java
@@ -27,7 +27,6 @@ import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentTransaction;
-import android.support.v4.app.ListFragment;
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.ActionBarDrawerToggle;
@@ -41,6 +40,9 @@ import android.widget.ListView;
 import org.adaway.R;
 import org.adaway.helper.OpenHelper;
 import org.adaway.helper.ResultHelper;
+import org.adaway.ui.hosts.HostsSourcesFragment;
+import org.adaway.ui.lists.ListsFragment;
+import org.adaway.ui.prefs.PrefsFragment;
 import org.adaway.util.Constants;
 import org.adaway.util.Log;
